Added score functionality
Added collision resolutions between player and pickups. Added a tracker for all of the pickups in the scene, as a means of scalability (currently planning on using minimal pickup instances, but if many were necessary, this tracker will handle them). The tracker is in charge of pickup event handling, subscribing and triggering a delegate. The score increase is then passed to a score handler, which uses a coroutine to smoothly change the score UI.