Added walls, wall scripting, and wall base
Added basic walls. Walls will fade out as the camera moves behind them in awkward positions (such as in corners, where the walls will block the camera) using a script. Attempted to use a shader for this but unfamiliar with Unity shader language and for some reason the code won't manipulate the alpha values - may revisit this. Also added flooring to the walls, to make it slightly more presentable upon fading out.