Added WatchTower and basic pickup spawning
Added a WatchTower model with full textures. Added a grass texture to the floor. Added script so that after a given time, the watchtower will spawn a pickup from those already instantiated and collected. Also added a progress bar on watchtowers to see how close they are on spawning a new pickup