Added building mechanic
Edited watchTower model to face the correct direction. Implemented a building mechanic. Players can now select a keypad 1-4 (this is actually just 1, as the full system isn't implemented)  or gamepad d-pad direction to select a building type. They can then use the mouse / right stick to see a preview of where the building will be placed. This building is restricted within a specified distance of the player, and will rotate towards the player. Players can then left click, or press the south gamepad button, to place the building, where it will begin its routines. Also edited watchTower code to rotate the progress bar towards the camera, and to only begin making pickups when placed.